Start redis from the project
Any dependencies to the project should be handled from the project. Unless otherwise stated.

So, `node-foreman` that is a clone of the ruby gem `foreman` is an awesome tool to start
servers and tools that need to run in development mode for an app to work.

Usage:

```sh
./node_modules/.bin/nf start
```

This will run the commands in the `Procfile`. If you install foreman globally too, you can
invoke with just `nf start`. But again, all dependencied for the app should be in the app,
so we add a development dependency.

About the Procfile: Form those that run a redis server as a deamon already, the fist command
just pings any existing and running server to check for existance. If not answering correctly
a new server is started.